requirements.txt
Django~=5.2.3
django-cors-headers==4.7.0
djangorestframework~=3.16.0
dj-database-url==3.0.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
